HTTP2 DTP
Summary: Use HTTP/2 to read block. HTTP/2 is a multiplex protocol so we do not need to open new connections to DN if there are multiple concurrent request. There are performance downgrading. One problem is that netty will use a separated thread to read write socket at client side, so there are extra context-switches. And another problem is netty HTTP/2 does not support writing FileRegion so we can not write large chunks which leads to a lower performance when reading chunks larger than 32K. Will optimize later.

Test Plan: UT and performance test.

Reviewers: zhouyingchao, zhangchen5, xiegang

Reviewed By: zhouyingchao

Subscribers: infra-dev

Differential Revision: https://phabricator.d.xiaomi.net/D62614